MATS/requirements.txt
# MATS 10.0: Sycophancy Entropy Control
# =====================================
# Dependencies for TransformerLens-based research

# Core
transformer_lens>=1.14.0
torch>=2.0.0
transformers>=4.36.0

# Data
datasets>=2.14.0

# Analysis
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
tqdm>=4.65.0
pyyaml>=6.0
einops>=0.7.0

# Reproducibility
gitpython>=3.1.0
